02-servicebus/requirements.txt
azure-servicebus
applicationinsights
openai
asyncio
opencensus-ext-azure